resize-image-function/requirements.txt
google-cloud-storage
pillow
